fix(Services): Separate distribution services
Some services are not loaded when installing the module.
