+++ config/services/distribution.yml
@@ -23,4 +23,11 @@ services:
     arguments:
       - "@mbo.distribution_api.client.http"
       - '@mbo.doctrine.cache.provider'
+
+  # Distribution API Client with addons user
+  mbo.cdc.client.connected_distribution_api:
+    class: PrestaShop\Module\Mbo\Distribution\ConnectedClient
+    arguments:
+      - "@mbo.distribution_api.client.http"
+      - '@mbo.doctrine.cache.provider'
       - '@=service("mbo.addons.user.provider").getUser()'
